# Conflicts:
#   src/main/java/de/htwberlin/casino/blackjack/adapter/out/persistence/GameJpaEntity.java
#   src/main/java/de/htwberlin/casino/blackjack/adapter/out/persistence/JpaRepositoryAdapter.java

+++ src/main/java/de/htwberlin/casino/blackjack/application/port/in/calculateChances/CalculateChancesCommand.java
@@ -1,4 +1,4 @@
 package de.htwberlin.casino.blackjack.application.port.in.calculateChances;
 
-public record CalculateChancesCommand(int gameId) {
+public record CalculateChancesCommand(Long gameId) {
 }
